Made contract test not depend on other tests results.
commit-id:a9594f70
