+++ crates/starknet/src/lib.rs
@@ -3,3 +3,6 @@ pub mod casm_contract_class;
 pub mod contract;
 pub mod contract_class;
 pub mod plugin;
+
+#[cfg(test)]
+pub mod test_utils;
